+++ release_notes.md
@@ -1,1 +1,3 @@
 ### Release notes
+
+- Add a new `VirtualHost` parameter to `[RabbitMQTrigger]`, `[RabbitMQ]` and `RabbitMQOptions` to allow to choose which [Virtual Hosts](https://www.rabbitmq.com/vhosts.html) to use when using host/user/password authentication method.